Initial pass at refactoring code
