+++ nginx/conf.d/default.conf
@@ -1,11 +1,19 @@
 # https://github.com/openresty/openresty/#resolvconf-parsing
 resolver local=on;
 
+# Get all the debugging
+error_log stderr notice;
+
+# Add multiple processes to test the eventual prod usecase
+#worker_processes auto;
 # cache for JWT verification results
 lua_shared_dict jwt_verification 10m;
 
 server {
-    listen 8080;
+    # reuseport will help scalability
+    # https://stackoverflow.com/questions/57494287/nginx-force-nginx-to-use-all-workers-for-load-balance
+    # https://blog.cloudflare.com/the-sad-state-of-linux-socket-balancing/
+    listen 8080 reuseport;
     server_name localhost;
 
     location / {
@@ -27,7 +35,7 @@ server {
         output_buffers 1 128k;
         # aio on; directio 512;
         client_max_body_size 20G;
-
+        content_by_lua_file /etc/nginx/lua/webdav_content.lua;
         access_by_lua_file /etc/nginx/lua/webdav_access.lua;
     }
 }
